BUGFIX: Check that a series is watched/unwatched
Series weren't checking to see if they were watched/unwatched.  This led
to a situation where a user could choose "Unwatched" and a season would
be picked with no Unwatched episodes in it.  Have also added in Dialogs
which exit the script early if the user chooses a course of action which
leads to an empty dialog box (i.e. choose "Unwatched" and there are no
unwatched shows.  If they then pick "Genre" they will end up with an
empty list - we don't want that)